Crate syntastica_queries

syntastica-queries

Collection of tree-sitter queries for syntastica.

See the project overview for more information.

This crate defines constants for three types of tree-sitter queries for lots of parsers. It is intended to be used via syntastica-parsers, syntastica-parsers-git, or syntastica-parsers-gitdep.

The three types of queries are:

highlights: defining the highlight captures for nodes

injections: defining where other languages are injected for highlighting

locals: keeping track of scopes, variables, parameters, etc. to have occurrences of those be highlighted the same everywhere

The constants are defined as <language_name>_<kind>[_CRATES_IO] where <kind> is one of HIGHLIGHTS, INJECTIONS, or LOCALS. The INJECTIONS and LOCALS may be empty for some languages. The constants with the _CRATES_IO suffix aim to be compatible with the latest version of the parser that was published on crates.io. These are used by syntastica-parsers, whereas the normal queries are used by syntastica-parsers-git and syntastica-parsers-gitdep.

The crate source is automatically generated with cargo xtask codegen inside the syntastica workspace.

Constants

ASM_HIGHLIGHTS

ASM_HIGHLIGHTS_CRATES_IO

ASM_INJECTIONS

ASM_INJECTIONS_CRATES_IO

ASM_LOCALS

ASM_LOCALS_CRATES_IO

BASH_HIGHLIGHTS

BASH_HIGHLIGHTS_CRATES_IO

BASH_INJECTIONS

BASH_INJECTIONS_CRATES_IO

BASH_LOCALS

BASH_LOCALS_CRATES_IO

CPP_HIGHLIGHTS

CPP_HIGHLIGHTS_CRATES_IO

CPP_INJECTIONS

CPP_INJECTIONS_CRATES_IO

CPP_LOCALS

CPP_LOCALS_CRATES_IO

CSS_HIGHLIGHTS

CSS_HIGHLIGHTS_CRATES_IO

CSS_INJECTIONS

CSS_INJECTIONS_CRATES_IO

CSS_LOCALS

CSS_LOCALS_CRATES_IO

C_HIGHLIGHTS

C_HIGHLIGHTS_CRATES_IO

C_INJECTIONS

C_INJECTIONS_CRATES_IO

C_LOCALS

C_LOCALS_CRATES_IO

GO_HIGHLIGHTS

GO_HIGHLIGHTS_CRATES_IO

GO_INJECTIONS

GO_INJECTIONS_CRATES_IO

GO_LOCALS

GO_LOCALS_CRATES_IO

HTML_HIGHLIGHTS

HTML_HIGHLIGHTS_CRATES_IO

HTML_INJECTIONS

HTML_INJECTIONS_CRATES_IO

HTML_LOCALS

HTML_LOCALS_CRATES_IO

JAVASCRIPT_HIGHLIGHTS

JAVASCRIPT_HIGHLIGHTS_CRATES_IO

JAVASCRIPT_INJECTIONS

JAVASCRIPT_INJECTIONS_CRATES_IO

JAVASCRIPT_LOCALS

JAVASCRIPT_LOCALS_CRATES_IO

JAVA_HIGHLIGHTS

JAVA_HIGHLIGHTS_CRATES_IO

JAVA_INJECTIONS

JAVA_INJECTIONS_CRATES_IO

JAVA_LOCALS

JAVA_LOCALS_CRATES_IO

JSON_HIGHLIGHTS

JSON_HIGHLIGHTS_CRATES_IO

JSON_INJECTIONS

JSON_INJECTIONS_CRATES_IO

JSON_LOCALS

JSON_LOCALS_CRATES_IO

PYTHON_HIGHLIGHTS

PYTHON_HIGHLIGHTS_CRATES_IO

PYTHON_INJECTIONS

PYTHON_INJECTIONS_CRATES_IO

PYTHON_LOCALS

PYTHON_LOCALS_CRATES_IO

REGEX_HIGHLIGHTS

REGEX_HIGHLIGHTS_CRATES_IO

REGEX_INJECTIONS

REGEX_INJECTIONS_CRATES_IO

REGEX_LOCALS

REGEX_LOCALS_CRATES_IO

RUST_HIGHLIGHTS

RUST_HIGHLIGHTS_CRATES_IO

RUST_INJECTIONS

RUST_INJECTIONS_CRATES_IO

RUST_LOCALS

RUST_LOCALS_CRATES_IO

TSX_HIGHLIGHTS

TSX_HIGHLIGHTS_CRATES_IO

TSX_INJECTIONS

TSX_INJECTIONS_CRATES_IO

TSX_LOCALS

TSX_LOCALS_CRATES_IO

TYPESCRIPT_HIGHLIGHTS

TYPESCRIPT_HIGHLIGHTS_CRATES_IO

TYPESCRIPT_INJECTIONS

TYPESCRIPT_INJECTIONS_CRATES_IO

TYPESCRIPT_LOCALS

TYPESCRIPT_LOCALS_CRATES_IO