Crate syntastica_queries

syntastica-queries

Collection of tree-sitter queries for syntastica.

See the project overview for more information.

This crate defines constants for three types of tree-sitter queries for lots of parsers. It is intended to be used via syntastica-parsers, syntastica-parsers-git, or syntastica-parsers-gitdep.

The three types of queries are:

highlights: defining the highlight captures for nodes

injections: defining where other languages are injected for highlighting

locals: keeping track of scopes, variables, parameters, etc. to have occurrences of those be highlighted the same everywhere

The constants are defined as <language_name>_<kind>[_CRATES_IO] where <kind> is one of HIGHLIGHTS, INJECTIONS, or LOCALS. The INJECTIONS and LOCALS may be empty for some languages. The constants with the _CRATES_IO suffix aim to be compatible with the latest version of the parser that was published on crates.io. These are used by syntastica-parsers, whereas the normal queries are used by syntastica-parsers-git and syntastica-parsers-gitdep.

The crate source is automatically generated with cargo xtask codegen inside the syntastica workspace.

Constants

ASM_HIGHLIGHTS

ASM_HIGHLIGHTS_CRATES_IO

ASM_INJECTIONS

ASM_INJECTIONS_CRATES_IO

ASM_LOCALS

ASM_LOCALS_CRATES_IO

BASH_HIGHLIGHTS

BASH_HIGHLIGHTS_CRATES_IO

BASH_INJECTIONS

BASH_INJECTIONS_CRATES_IO

BASH_LOCALS

BASH_LOCALS_CRATES_IO

BIBTEX_HIGHLIGHTS

BIBTEX_HIGHLIGHTS_CRATES_IO

BIBTEX_INJECTIONS

BIBTEX_INJECTIONS_CRATES_IO

BIBTEX_LOCALS

BIBTEX_LOCALS_CRATES_IO

COMMENT_HIGHLIGHTS

COMMENT_HIGHLIGHTS_CRATES_IO

COMMENT_INJECTIONS

COMMENT_INJECTIONS_CRATES_IO

COMMENT_LOCALS

COMMENT_LOCALS_CRATES_IO

CPP_HIGHLIGHTS

CPP_HIGHLIGHTS_CRATES_IO

CPP_INJECTIONS

CPP_INJECTIONS_CRATES_IO

CPP_LOCALS

CPP_LOCALS_CRATES_IO

CSS_HIGHLIGHTS

CSS_HIGHLIGHTS_CRATES_IO

CSS_INJECTIONS

CSS_INJECTIONS_CRATES_IO

CSS_LOCALS

CSS_LOCALS_CRATES_IO

C_HIGHLIGHTS

C_HIGHLIGHTS_CRATES_IO

C_INJECTIONS

C_INJECTIONS_CRATES_IO

C_LOCALS

C_LOCALS_CRATES_IO

C_SHARP_HIGHLIGHTS

C_SHARP_HIGHLIGHTS_CRATES_IO

C_SHARP_INJECTIONS

C_SHARP_INJECTIONS_CRATES_IO

C_SHARP_LOCALS

C_SHARP_LOCALS_CRATES_IO

DART_HIGHLIGHTS

DART_HIGHLIGHTS_CRATES_IO

DART_INJECTIONS

DART_INJECTIONS_CRATES_IO

DART_LOCALS

DART_LOCALS_CRATES_IO

DIFF_HIGHLIGHTS

DIFF_HIGHLIGHTS_CRATES_IO

DIFF_INJECTIONS

DIFF_INJECTIONS_CRATES_IO

DIFF_LOCALS

DIFF_LOCALS_CRATES_IO

DOCKERFILE_HIGHLIGHTS

DOCKERFILE_HIGHLIGHTS_CRATES_IO

DOCKERFILE_INJECTIONS

DOCKERFILE_INJECTIONS_CRATES_IO

DOCKERFILE_LOCALS

DOCKERFILE_LOCALS_CRATES_IO

EBNF_HIGHLIGHTS

EBNF_HIGHLIGHTS_CRATES_IO

EBNF_INJECTIONS

EBNF_INJECTIONS_CRATES_IO

EBNF_LOCALS

EBNF_LOCALS_CRATES_IO

EJS_HIGHLIGHTS

EJS_HIGHLIGHTS_CRATES_IO

EJS_INJECTIONS

EJS_INJECTIONS_CRATES_IO

EJS_LOCALS

EJS_LOCALS_CRATES_IO

ELIXIR_HIGHLIGHTS

ELIXIR_HIGHLIGHTS_CRATES_IO

ELIXIR_INJECTIONS

ELIXIR_INJECTIONS_CRATES_IO

ELIXIR_LOCALS

ELIXIR_LOCALS_CRATES_IO

ERB_HIGHLIGHTS

ERB_HIGHLIGHTS_CRATES_IO

ERB_INJECTIONS

ERB_INJECTIONS_CRATES_IO

ERB_LOCALS

ERB_LOCALS_CRATES_IO

FISH_HIGHLIGHTS

FISH_HIGHLIGHTS_CRATES_IO

FISH_INJECTIONS

FISH_INJECTIONS_CRATES_IO

FISH_LOCALS

FISH_LOCALS_CRATES_IO

GO_HIGHLIGHTS

GO_HIGHLIGHTS_CRATES_IO

GO_INJECTIONS

GO_INJECTIONS_CRATES_IO

GO_LOCALS

GO_LOCALS_CRATES_IO

HASKELL_HIGHLIGHTS

HASKELL_HIGHLIGHTS_CRATES_IO

HASKELL_INJECTIONS

HASKELL_INJECTIONS_CRATES_IO

HASKELL_LOCALS

HASKELL_LOCALS_CRATES_IO

HEXDUMP_HIGHLIGHTS

HEXDUMP_HIGHLIGHTS_CRATES_IO

HEXDUMP_INJECTIONS

HEXDUMP_INJECTIONS_CRATES_IO

HEXDUMP_LOCALS

HEXDUMP_LOCALS_CRATES_IO

HTML_HIGHLIGHTS

HTML_HIGHLIGHTS_CRATES_IO

HTML_INJECTIONS

HTML_INJECTIONS_CRATES_IO

HTML_LOCALS

HTML_LOCALS_CRATES_IO

JAVASCRIPT_HIGHLIGHTS

JAVASCRIPT_HIGHLIGHTS_CRATES_IO

JAVASCRIPT_INJECTIONS

JAVASCRIPT_INJECTIONS_CRATES_IO

JAVASCRIPT_LOCALS

JAVASCRIPT_LOCALS_CRATES_IO

JAVA_HIGHLIGHTS

JAVA_HIGHLIGHTS_CRATES_IO

JAVA_INJECTIONS

JAVA_INJECTIONS_CRATES_IO

JAVA_LOCALS

JAVA_LOCALS_CRATES_IO

JSDOC_HIGHLIGHTS

JSDOC_HIGHLIGHTS_CRATES_IO

JSDOC_INJECTIONS

JSDOC_INJECTIONS_CRATES_IO

JSDOC_LOCALS

JSDOC_LOCALS_CRATES_IO

JSON5_HIGHLIGHTS

JSON5_HIGHLIGHTS_CRATES_IO

JSON5_INJECTIONS

JSON5_INJECTIONS_CRATES_IO

JSON5_LOCALS

JSON5_LOCALS_CRATES_IO

JSONC_HIGHLIGHTS

JSONC_HIGHLIGHTS_CRATES_IO

JSONC_INJECTIONS

JSONC_INJECTIONS_CRATES_IO

JSONC_LOCALS

JSONC_LOCALS_CRATES_IO

JSON_HIGHLIGHTS

JSON_HIGHLIGHTS_CRATES_IO

JSON_INJECTIONS

JSON_INJECTIONS_CRATES_IO

JSON_LOCALS

JSON_LOCALS_CRATES_IO

JULIA_HIGHLIGHTS

JULIA_HIGHLIGHTS_CRATES_IO

JULIA_INJECTIONS

JULIA_INJECTIONS_CRATES_IO

JULIA_LOCALS

JULIA_LOCALS_CRATES_IO

KOTLIN_HIGHLIGHTS

KOTLIN_HIGHLIGHTS_CRATES_IO

KOTLIN_INJECTIONS

KOTLIN_INJECTIONS_CRATES_IO

KOTLIN_LOCALS

KOTLIN_LOCALS_CRATES_IO

LALRPOP_HIGHLIGHTS

LALRPOP_HIGHLIGHTS_CRATES_IO

LALRPOP_INJECTIONS

LALRPOP_INJECTIONS_CRATES_IO

LALRPOP_LOCALS

LALRPOP_LOCALS_CRATES_IO

LATEX_HIGHLIGHTS

LATEX_HIGHLIGHTS_CRATES_IO

LATEX_INJECTIONS

LATEX_INJECTIONS_CRATES_IO

LATEX_LOCALS

LATEX_LOCALS_CRATES_IO

LLVM_HIGHLIGHTS

LLVM_HIGHLIGHTS_CRATES_IO

LLVM_INJECTIONS

LLVM_INJECTIONS_CRATES_IO

LLVM_LOCALS

LLVM_LOCALS_CRATES_IO

LUA_HIGHLIGHTS

LUA_HIGHLIGHTS_CRATES_IO

LUA_INJECTIONS

LUA_INJECTIONS_CRATES_IO

LUA_LOCALS

LUA_LOCALS_CRATES_IO

MARKDOWN_HIGHLIGHTS

MARKDOWN_HIGHLIGHTS_CRATES_IO

MARKDOWN_INJECTIONS

MARKDOWN_INJECTIONS_CRATES_IO

MARKDOWN_INLINE_HIGHLIGHTS

MARKDOWN_INLINE_HIGHLIGHTS_CRATES_IO

MARKDOWN_INLINE_INJECTIONS

MARKDOWN_INLINE_INJECTIONS_CRATES_IO

MARKDOWN_INLINE_LOCALS

MARKDOWN_INLINE_LOCALS_CRATES_IO

MARKDOWN_LOCALS

MARKDOWN_LOCALS_CRATES_IO

OCAML_HIGHLIGHTS

OCAML_HIGHLIGHTS_CRATES_IO

OCAML_INJECTIONS

OCAML_INJECTIONS_CRATES_IO

OCAML_INTERFACE_HIGHLIGHTS

OCAML_INTERFACE_HIGHLIGHTS_CRATES_IO

OCAML_INTERFACE_INJECTIONS

OCAML_INTERFACE_INJECTIONS_CRATES_IO

OCAML_INTERFACE_LOCALS

OCAML_INTERFACE_LOCALS_CRATES_IO

OCAML_LOCALS

OCAML_LOCALS_CRATES_IO

PHP_HIGHLIGHTS

PHP_HIGHLIGHTS_CRATES_IO

PHP_INJECTIONS

PHP_INJECTIONS_CRATES_IO

PHP_LOCALS

PHP_LOCALS_CRATES_IO

PRINTF_HIGHLIGHTS

PRINTF_HIGHLIGHTS_CRATES_IO

PRINTF_INJECTIONS

PRINTF_INJECTIONS_CRATES_IO

PRINTF_LOCALS

PRINTF_LOCALS_CRATES_IO

PYTHON_HIGHLIGHTS

PYTHON_HIGHLIGHTS_CRATES_IO

PYTHON_INJECTIONS

PYTHON_INJECTIONS_CRATES_IO

PYTHON_LOCALS

PYTHON_LOCALS_CRATES_IO

QL_HIGHLIGHTS

QL_HIGHLIGHTS_CRATES_IO

QL_INJECTIONS

QL_INJECTIONS_CRATES_IO

QL_LOCALS

QL_LOCALS_CRATES_IO

REGEX_HIGHLIGHTS

REGEX_HIGHLIGHTS_CRATES_IO

REGEX_INJECTIONS

REGEX_INJECTIONS_CRATES_IO

REGEX_LOCALS

REGEX_LOCALS_CRATES_IO

RUBY_HIGHLIGHTS

RUBY_HIGHLIGHTS_CRATES_IO

RUBY_INJECTIONS

RUBY_INJECTIONS_CRATES_IO

RUBY_LOCALS

RUBY_LOCALS_CRATES_IO

RUSH_HIGHLIGHTS

RUSH_HIGHLIGHTS_CRATES_IO

RUSH_INJECTIONS

RUSH_INJECTIONS_CRATES_IO

RUSH_LOCALS

RUSH_LOCALS_CRATES_IO

RUST_HIGHLIGHTS

RUST_HIGHLIGHTS_CRATES_IO

RUST_INJECTIONS

RUST_INJECTIONS_CRATES_IO

RUST_LOCALS

RUST_LOCALS_CRATES_IO

SCALA_HIGHLIGHTS

SCALA_HIGHLIGHTS_CRATES_IO

SCALA_INJECTIONS

SCALA_INJECTIONS_CRATES_IO

SCALA_LOCALS

SCALA_LOCALS_CRATES_IO

SCSS_HIGHLIGHTS

SCSS_HIGHLIGHTS_CRATES_IO

SCSS_INJECTIONS

SCSS_INJECTIONS_CRATES_IO

SCSS_LOCALS

SCSS_LOCALS_CRATES_IO

SQL_HIGHLIGHTS

SQL_HIGHLIGHTS_CRATES_IO

SQL_INJECTIONS

SQL_INJECTIONS_CRATES_IO

SQL_LOCALS

SQL_LOCALS_CRATES_IO

TOML_HIGHLIGHTS

TOML_HIGHLIGHTS_CRATES_IO

TOML_INJECTIONS

TOML_INJECTIONS_CRATES_IO

TOML_LOCALS

TOML_LOCALS_CRATES_IO

TSX_HIGHLIGHTS

TSX_HIGHLIGHTS_CRATES_IO

TSX_INJECTIONS

TSX_INJECTIONS_CRATES_IO

TSX_LOCALS

TSX_LOCALS_CRATES_IO

TYPESCRIPT_HIGHLIGHTS

TYPESCRIPT_HIGHLIGHTS_CRATES_IO

TYPESCRIPT_INJECTIONS

TYPESCRIPT_INJECTIONS_CRATES_IO

TYPESCRIPT_LOCALS

TYPESCRIPT_LOCALS_CRATES_IO

TYPST_HIGHLIGHTS

TYPST_HIGHLIGHTS_CRATES_IO

TYPST_INJECTIONS

TYPST_INJECTIONS_CRATES_IO

TYPST_LOCALS

TYPST_LOCALS_CRATES_IO

URSA_HIGHLIGHTS

URSA_HIGHLIGHTS_CRATES_IO

URSA_INJECTIONS

URSA_INJECTIONS_CRATES_IO

URSA_LOCALS

URSA_LOCALS_CRATES_IO

VERILOG_HIGHLIGHTS

VERILOG_HIGHLIGHTS_CRATES_IO

VERILOG_INJECTIONS

VERILOG_INJECTIONS_CRATES_IO

VERILOG_LOCALS

VERILOG_LOCALS_CRATES_IO

WAT_HIGHLIGHTS

WAT_HIGHLIGHTS_CRATES_IO

WAT_INJECTIONS

WAT_INJECTIONS_CRATES_IO

WAT_LOCALS

WAT_LOCALS_CRATES_IO

YAML_HIGHLIGHTS

YAML_HIGHLIGHTS_CRATES_IO

YAML_INJECTIONS

YAML_INJECTIONS_CRATES_IO

YAML_LOCALS

YAML_LOCALS_CRATES_IO

ZIG_HIGHLIGHTS

ZIG_HIGHLIGHTS_CRATES_IO

ZIG_INJECTIONS

ZIG_INJECTIONS_CRATES_IO

ZIG_LOCALS

ZIG_LOCALS_CRATES_IO